Creador de resetPass

+++ src/main/java/com/linkhogar/application/user/deleteHouseFavourite/DeleteHouseFAvouriteCommandHandler.java
@@ -1,10 +1,13 @@
 package com.linkhogar.application.user.deleteHouseFavourite;
 
 import com.linkhogar.domain.common.result.Result;
+import com.linkhogar.domain.user.HouseFavourite;
 import com.linkhogar.domain.user.HouseFavouriteErrors;
 import com.linkhogar.domain.user.HouseFavouriteRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -13,13 +16,16 @@ public class DeleteHouseFAvouriteCommandHandler {
 
     public Result<Void> handler(DeleteHouseFavouriteCommand command){
         try {
+            Optional<HouseFavourite> houseFavouriteOptional = houseFavouriteRepository.getFavourite(command.userId(), command.houseId());
+
+            if(houseFavouriteOptional.isEmpty()){return Result.success(null);}
+
             // Asumiendo que tu comando tiene los ID necesarios
             houseFavouriteRepository.deleteFavourite(
                     command.userId(),
                     command.houseId()
             );
 
-            // Retornamos éxito (ajusta al método exacto de tu clase Result)
             return Result.success(null);
 
         } catch (Exception e) {
